move almost all views to separate packages
