+++ contributors/views/generic_list_views/leaderboards.py
@@ -1,7 +1,7 @@
 from django.views import generic
 
 from contributors.models import Contributor
-from contributors.views.mixins import (
+from contributors.views.utils.mixins import (
     LeaderboardQueryMixin,
     TableSortSearchAndPaginationMixin,
 )
