+++ contributors/views/organizations_views/organizations.py
@@ -3,7 +3,7 @@ from django.utils.translation import gettext_lazy as _
 from django.views import generic
 
 from contributors.models import Organization
-from contributors.views.mixins import TableSortSearchAndPaginationMixin
+from contributors.views.utils.mixins import TableSortSearchAndPaginationMixin
 
 
 class ListView(TableSortSearchAndPaginationMixin, generic.ListView):
